Conga line 1, timed out for some cases.
First submision of the conga line problem. It timed out for some test
cases and didn't seem obviously incorrect other than that. It might be a
fairly simple fix, as I am not using fast input/output methods, or there
might be something deeper going on in the code.

Either way, an interesting software engineering problem!