Industrial Spy: accepted submission.
Final imlementation of Miller-Rabin algorithm on all possible numbers
for a list of up to seven digits. Getting length r permutations was a
little finicky, as well as lower bounds (is 2 prime? etc). But overall a
really cool problem and exposure to some of the pieces of RSA.

It's amazing an integer n can be prime tested in O(lg lg lg (n)). Good
problem.