Mate in One: Partially Correct.
This might need re-done with better software engineering applied. The
first 6 (and probably visible) test cases are passing, but the 7th test
case is failing. I'm having a hard time seeing where it could be wrong.
The authors desire to make a point, IE, that it's hard to track down
bugs in _very complicated_ code has been made. Managing complexity is
the name of the game.

Very frustrating. Another case of where you need to understand the
problem definition *exactly*. It's hard to pay enough attention to
understand *every single small distinct possibility* in a chess game.
Frustrating problem, hopefully I get to come back to it with better
software engineering!